xhci: TD-fragment, align the unsplittable case with a bounce buffer
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Mathias Nyman <mathias.nyman@linux.intel.com>
commit f9c589e142d04b8a19eb382162f804d17102b5ed

If the last trb before a link is not packet size aligned, and is not
splittable then use a bounce buffer for that chunk of max packet size
unalignable data.

Allocate a max packet size bounce buffer for every segment of a bulk
endpoint ring at the same time as allocating the ring.
If we need to align the data before the link trb in that segment then
copy the data to the segment bounce buffer, dma map it, and enqueue it.
Once the td finishes, or is cancelled, unmap it.

For in transfers we need to first map the bounce buffer, then queue it,
after it finishes, copy the bounce buffer to the original sg list, and
finally unmap it

	Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit f9c589e142d04b8a19eb382162f804d17102b5ed)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>