usb: hub: Wait for connection to be reestablished after port reset
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
Rebuild_CHGLOG: - [usb] hub: Wait for connection to be reestablished after port reset (Torez Smith) [1383823]
Rebuild_FUZZ: 96.06%
commit-author Guenter Roeck <linux@roeck-us.net>
commit 22547c4cc4fe20698a6a85a55b8788859134b8e4

On a system with a defective USB device connected to an USB hub,
an endless sequence of port connect events was observed. The sequence
of events as observed is as follows:

- Port reports connected event (port status=USB_PORT_STAT_CONNECTION).
- Event handler debounces port and resets it by calling hub_port_reset().
- hub_port_reset() calls hub_port_wait_reset() to wait for the reset
  to complete.
- The reset completes, but USB_PORT_STAT_CONNECTION is not immediately
  set in the port status register.
- hub_port_wait_reset() returns -ENOTCONN.
- Port initialization sequence is aborted.
- A few milliseconds later, the port again reports a connected event,
  and the sequence repeats.

This continues either forever or, randomly, stops if the connection
is already re-established when the port status is read. It results in
a high rate of udev events. This in turn destabilizes userspace since
the above sequence holds the device mutex pretty much continuously
and prevents userspace from actually reading the device status.

To prevent the problem from happening, let's wait for the connection
to be re-established after a port reset. If the device was actually
disconnected, the code will still return an error, but it will do so
only after the long reset timeout.

	Cc: Douglas Anderson <dianders@chromium.org>
	Signed-off-by: Guenter Roeck <linux@roeck-us.net>
	Acked-by: Alan Stern <stern@rowland.harvard.edu>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 22547c4cc4fe20698a6a85a55b8788859134b8e4)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>