fsl/usb: Workarourd for USB erratum-A005697
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Changming Huang <jerry.huang@nxp.com>
commit 9d4b82706357f2eb23f45309227fc94d11eea255
Empty-Commit: Cherry-Pick Conflicts during history rebuild.
Will be included in final tarball splat. Ref for failed cherry-pick at:
ciq/ciq_backports/kernel-3.10.0-693.el7/9d4b8270.failed

The EHCI specification states the following in the SUSP bit description:
In the Suspend state, the port is sensitive to resume detection.
Note that the bit status does not change until the port is suspended and
that there may be a delay in suspending a port if there is a transaction
currently in progress on the USB.

However, in NXP USBDR controller, the PORTSCx[SUSP] bit changes immediately
when the application sets it and not when the port is actually suspended.

So the application must wait for at least 10 milliseconds after a port
indicates that it is suspended, to make sure this port has entered
suspended state before initiating this port resume using the Force Port
Resume bit. This bit is for NXP controller, not EHCI compatible.

	Signed-off-by: Changming Huang <jerry.huang@nxp.com>
	Signed-off-by: Ramneek Mehresh <ramneek.mehresh@nxp.com>
	Acked-by: Alan Stern <stern@rowland.harvard.edu>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 9d4b82706357f2eb23f45309227fc94d11eea255)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>

# Conflicts:
#	drivers/usb/host/ehci-fsl.c
#	drivers/usb/host/fsl-mph-dr-of.c