net/mlx5: Group similer rules under the same fte
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
Rebuild_CHGLOG: - [netdrv] mlx5: Group similer rules under the same fte (Don Dutile) [1385330 1417286]
Rebuild_FUZZ: 95.65%
commit-author Mark Bloch <markb@mellanox.com>
commit a62249857779733dc7cb5d43d6ecdc35fa81b88f

When adding a new rule, if we can match it with compare_match_value and
flow tag we might be able to insert the rule to the same fte.
In order to do that, there must be an overlap between the actions of the
fte and the new rule.

When updating the action of an existing fte, we must tell the firmware
we are doing so.

	Signed-off-by: Mark Bloch <markb@mellanox.com>
	Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>
	Signed-off-by: Leon Romanovsky <leon@kernel.org>
(cherry picked from commit a62249857779733dc7cb5d43d6ecdc35fa81b88f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>