IB/uverbs: Add support for Vxlan protocol
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Moses Reuben <mosesr@mellanox.com>
commit a0cb4c759af12943806564294fa53ab08cb7cf93

Add ib_uverbs_flow_spec_tunnel to define the rule to match Vxlan,
the type, size, reserved fields are identical to rest of the protocols,
and are used to identify the spec.
The tunnel id is the vni value of the Vxlan protocol, and it is used
as part of the steering rule, it is limited by the mask.
The steering rule configured on the hardware does a match
according to vni and other protocols.
In the same way as rest of the protocols that we match,
the uniq field's of each protocol are represented on
the val and the mask.

	Signed-off-by: Moses Reuben <mosesr@mellanox.com>
	Reviewed-by: Maor Gottlieb <maorg@mellanox.com>
	Signed-off-by: Leon Romanovsky <leon@kernel.org>
	Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit a0cb4c759af12943806564294fa53ab08cb7cf93)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>