net/mlx4_core: Fix when to save some qp context flags for dynamic VST to VGT transitions
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
Rebuild_CHGLOG: - [netdrv] mlx4_core: Fix when to save some qp context flags for dynamic VST to VGT transitions (Don Dutile) [1385329 1417286 1429188]
Rebuild_FUZZ: 97.67%
commit-author Jack Morgenstein <jackm@dev.mellanox.co.il>
commit 7c3945bc2073554bb2ecf983e073dee686679c53

Save the qp context flags byte containing the flag disabling vlan stripping
in the RESET to INIT qp transition, rather than in the INIT to RTR
transition. Per the firmware spec, the flags in this byte are active
in the RESET to INIT transition.

As a result of saving the flags in the incorrect qp transition, when
switching dynamically from VGT to VST and back to VGT, the vlan
remained stripped (as is required for VST) and did not return to
not-stripped (as is required for VGT).

Fixes: f0f829bf42cd ("net/mlx4_core: Add immediate activate for VGT->VST->VGT")
	Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
	Signed-off-by: Tariq Toukan <tariqt@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 7c3945bc2073554bb2ecf983e073dee686679c53)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>