powerpc32: optimise a few instructions in csum_partial()
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
Rebuild_CHGLOG: - [powerpc] powerpc/32: optimise a few instructions in csum_partial() (Gustavo Duarte) [1430834]
Rebuild_FUZZ: 99.12%
commit-author Christophe Leroy <christophe.leroy@c-s.fr>
commit 48821a34b1bdc5d89505cb814b3f7c166940f200

r5 does contain the value to be updated, so lets use r5 all way long
for that. It makes the code more readable.

To avoid confusion, it is better to use adde instead of addc

The first addition is useless. Its only purpose is to clear carry.
As r4 is a signed int that is always positive, this can be done by
using srawi instead of srwi

Let's also remove the comment about bdnz having no overhead as it
is not correct on all powerpc, at least on MPC8xx

In the last part, in our situation, the remaining quantity of bytes
to be proceeded is between 0 and 3. Therefore, we can base that part
on the value of bit 31 and bit 30 of r4 instead of anding r4 with 3
then proceding on comparisons and substractions.

	Signed-off-by: Christophe Leroy <christophe.leroy@c-s.fr>
	Signed-off-by: Scott Wood <oss@buserror.net>
(cherry picked from commit 48821a34b1bdc5d89505cb814b3f7c166940f200)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>