NTB: Fix macro parameter conflict with field name
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
Rebuild_CHGLOG: - [ntb] Fix macro parameter conflict with field name (Suravee Suthikulpanit) [1303727]
Rebuild_FUZZ: 94.62%
commit-author Allen Hubbe <Allen.Hubbe@emc.com>
commit 03beaec80d57e7a540b1990188f8a97a076da6de
Empty-Commit: Cherry-Pick Conflicts during history rebuild.
Will be included in final tarball splat. Ref for failed cherry-pick at:
ciq/ciq_backports/kernel-3.10.0-693.el7/03beaec8.failed

If the parameter given to the macro is replaced throughout the macro as
it is evaluated.  The intent is that the macro parameter should replace
the only the first parameter to container_of().  However, the way the
macro was written, it would also inadvertantly replace a structure field
name.  If a parameter of any other name is given to the macro, it will
fail to compile, if the structure does not contain a field of the same
name.  At worst, it will compile, and hide improper access of an
unintended field in the structure.

Change the macro parameter name, so it does not conflict with the
structure field name.

	Signed-off-by: Allen Hubbe <Allen.Hubbe@emc.com>
	Acked-by: Dave Jiang <dave.jiang@intel.com>
	Signed-off-by: Jon Mason <jdmason@kudzu.us>
(cherry picked from commit 03beaec80d57e7a540b1990188f8a97a076da6de)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>

# Conflicts:
#	drivers/ntb/hw/intel/ntb_hw_intel.h