mnt: Prevent pivot_root from creating a loop in the mount tree
jira LE-1907
cve CVE-2014-7970
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Eric W. Biederman <ebiederm@xmission.com>
commit 0d0826019e529f21c84687521d03f60cd241ca7d

Andy Lutomirski recently demonstrated that when chroot is used to set
the root path below the path for the new ``root'' passed to pivot_root
the pivot_root system call succeeds and leaks mounts.

In examining the code I see that starting with a new root that is
below the current root in the mount tree will result in a loop in the
mount tree after the mounts are detached and then reattached to one
another.  Resulting in all kinds of ugliness including a leak of that
mounts involved in the leak of the mount loop.

Prevent this problem by ensuring that the new mount is reachable from
the current root of the mount tree.

[Added stable cc.  Fixes CVE-2014-7970.  --Andy]

	Cc: stable@vger.kernel.org
	Reported-by: Andy Lutomirski <luto@amacapital.net>
	Reviewed-by: Andy Lutomirski <luto@amacapital.net>
Link: http://lkml.kernel.org/r/87bnpmihks.fsf@x220.int.ebiederm.org
	Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
	Signed-off-by: Andy Lutomirski <luto@amacapital.net>
(cherry picked from commit 0d0826019e529f21c84687521d03f60cd241ca7d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>